requirements.txt
# The only version that works currently
# duckcli==0.0.1
# dbt-core==1.6.1
# Database adapter
dbt-duckdb==1.6.0
# dbt-postgres==1.5.2
# data-diff==0.7.13
# install master to get the postgres plugin working
# git+https://github.com/jwills/dbt-duckdb.git@master
